Moved the database structure and functions into its own folder
