+++ thinkback/database/database.py
@@ -1,0 +1,36 @@
+# thinkback/database.py
+
+from flask import g
+from flask import current_app as app
+import sqlite3
+
+
+def drop_db():
+    db = get_db()
+    with app.open_resource('drop.sql', mode='r') as f:
+        db.cursor().executescript(f.read())
+    db.commit()
+
+
+def init_db():
+    db = get_db()
+    with app.open_resource('schema.sql', mode='r') as f:
+        db.cursor().executescript(f.read())
+    db.commit()
+
+
+def get_db():
+    """Opens a new database connection if there is none yet for the
+        current application context.
+        """
+    if not hasattr(g, 'thinkback.db'):
+        g.sqlite_db = connect_db()
+    return g.sqlite_db
+
+
+def connect_db():
+    """Connects to the specific database."""
+    rv = sqlite3.connect(app.config['DATABASE'])
+    rv.row_factory = sqlite3.Row
+    return rv
+
