+++ thinkback/thinkback.py
@@ -3,6 +3,7 @@ from flask import Flask, render_template, request, session, g, redirect, url_for
 import os
 from .views import about, assignments, upload
 import sqlite3
+from .database import database
 
 UPLOAD_FOLDER = './uploads'
 SOLUTIONS_FOLDER = './impl'
@@ -33,45 +34,15 @@ app.config.from_envvar('THINKBACK_SETTINGS', silent=True)
 @app.cli.command('dropdb')
 def drop_db_command():
     print("Dropping database.")
-    drop_db()
+    database.drop_db()
     print("Dropped database.")
 
 
 @app.cli.command('initdb')
 def initdb_command():
     print("""Initializes the database.""")
-    init_db()
+    database.init_db()
     print('Initialized the database.')
-
-
-def drop_db():
-    db = get_db()
-    with app.open_resource('drop.sql', mode='r') as f:
-        db.cursor().executescript(f.read())
-    db.commit()
-
-
-def init_db():
-    db = get_db()
-    with app.open_resource('schema.sql', mode='r') as f:
-        db.cursor().executescript(f.read())
-    db.commit()
-
-
-def get_db():
-    """Opens a new database connection if there is none yet for the
-        current application context.
-        """
-    if not hasattr(g, 'thinkback.db'):
-        g.sqlite_db = connect_db()
-    return g.sqlite_db
-
-
-def connect_db():
-    """Connects to the specific database."""
-    rv = sqlite3.connect(app.config['DATABASE'])
-    rv.row_factory = sqlite3.Row
-    return rv
 
 
 @app.teardown_appcontext
@@ -79,7 +50,6 @@ def close_db(error):
     print("""Closes the database again at the end of the request.""")
     if hasattr(g, 'thinkback.db'):
         g.sqlite_db.close()
-
 
 @app.route('/index')
 @app.route('/')
